+++ storybook/webpack.config.js
@@ -1,33 +1,44 @@
-const findCacheDir = require('find-cache-dir');
-
 module.exports = {
   module: {
-    loaders: [
+    rules: [
       {
         test: /\.(?:png|svg)$/,
-        loader: 'url-loader',
+        use: 'url-loader',
       },
       {
         test: /\.jsx?$/,
         exclude: [/(node_modules)/, /react-css-themr/],
-        loader: require.resolve('babel-loader'),
-        query: {
-          babelrc: false,
-          cacheDirectory: findCacheDir({ name: 'react-storybook' }),
-          presets: ['env', 'react'],
-          plugins: [
-            'transform-flow-strip-types',
-            'transform-decorators-legacy',
-            'transform-class-properties',
-            'lodash',
-          ],
-        },
-      }, {
+        use: [{
+          loader: 'babel-loader',
+          options: {
+            babelrc: false,
+            presets: [
+              ['env', { "modules": false }],
+              'react'
+            ],
+            plugins: [
+              'transform-flow-strip-types',
+              'transform-decorators-legacy',
+              'transform-class-properties',
+              'lodash',
+            ],
+          }
+        }],
+      },
+      {
         test: /\.scss$/,
-        loaders: [
-          'style-loader?sourceMap',
-          'css-loader?sourceMap&modules&localIdentName=[name]_[local]&importLoaders=1',
-          'sass-loader?sourceMap',
+        use: [
+          { loader: 'style-loader', options: { sourceMap: true }},
+          {
+            loader: 'css-loader',
+            options: {
+              sourceMap: true,
+              modules: true,
+              localIdentName: '[name]_[local]',
+              importLoaders: true,
+            }
+          },
+          { loader: 'sass-loader', options: { sourceMap: true }},
         ]
       }
     ]